Optimized events pushing
This also resolves a memory leak that was occurring on staging, probably
due to an event handler never being removed from the event object
(disposed). Removed usage of the helper Request class, which is all but
deprecated. Now writes serialized content directly to the http stream
avoiding the large object heap if lots of events occur at the same time.