Now ignores created stamp on NetworkObject
NetworkObject only attribute was a timestamp for when the object was
created. This is useful through out some of action in Procon.Net and
potentially plugins, but since only events stream out of Procon it is
then stamped/wrapped by an event object, so any subsequent timestamps
are useless and spammy.