Ensured isolation within unit tests
The Procon.Core.Test.* was using shared static values between tests by
accident. These static values are now overwritten on every test. We
don't actually reference the static value but instead copy it to a
reference that exists in the test. We just weren't creating a new static
reference to begin with on each test. This does make the tests take a
little bit longer, but it's how it was supposed to run initially