Process is now terminated if service domain cannot be unloaded
Noticed strange errors occurring on staging where multiple versions of
the same packages were installed. As predicated in the comments for this
process, the issue was the domain failing to unload but the process
continuing to update packages. The packages couldn't be uninstalled
because the AppDomain that couldn't be unloaded still had locks on the
assemblies.