No longer saved traits or descriptions
Thinking on it, if we ever needed to change them they would be
overwritten on-load. This method means plugins are responsible for
setting up their permissions (traits and descriptions) on each load,
just like Procon is, however it also means we can make alterations in
the future easily.

Really it's also kind of pointless saving this data since it's hard
coded into Procon or the plugin.