Now fires events when a repository is added/removed
Though it is bound to two specific commands going in, instead of the
action of editing the variables. I'm unsure if the event firing needs to
be moved down the callstack to the source or if it makes more sense
firing these events on add/remove of he actual repository, but that may
be spammy.